--- models/engine/__init__.py
@@ -1,17 +1,0 @@
-#!/usr/bin/python3
-"""
-This module instantiates the type of storage that will be used for projects
-"""
-
-from os import getenv
-
-storage_type = getenv('HBNB_TYPE_STORAGE')
-
-if storage_type == 'db':
-    from models.engine.db_storage import DBStorage
-    storage = DBStorage()
-else:
-    from models.engine.file_storage import FileStorage
-    storage = FileStorage()
-
-storage. reload()
